infinite axis ticks

+++ src/main/java/com/duncpro/cartesianj/awt/UnitConverter.java
@@ -17,13 +17,13 @@ public abstract class UnitConverter {
     protected abstract int getViewDimensionPx(Axis axis);
 
     public int toPx(double n, Axis axis) {
-        double totalOnScreenTicks = getViewDimensionPx(axis) / plane.getTickWidth(axis);
+        double totalOnScreenTicks =  (double) getViewDimensionPx(axis) / plane.getTickWidth(axis);
         double pxPerTick = getViewDimensionPx(axis) / totalOnScreenTicks;
         return toIntExact(round((n / plane.getStepSize(axis)) * pxPerTick));
     }
 
     public double toStep(int px, Axis axis) {
-        double totalOnScreenTicks = getViewDimensionPx(axis) / plane.getTickWidth(axis);
+        double totalOnScreenTicks = (double) getViewDimensionPx(axis) / plane.getTickWidth(axis);
         double tickPerPx = totalOnScreenTicks / getViewDimensionPx(axis);
         return (tickPerPx * px) * plane.getStepSize(axis);
     }
